server/requirements.txt
Django==3.0.8
psycopg2-binary==2.8.5
djangorestframework==3.11.0
django-filter==2.3.0